feat(feature/router): add vue-router
