+++ package.json
@@ -11,7 +11,8 @@
   },
   "dependencies": {
     "json-server": "^1.0.0-beta.3",
-    "vue": "^3.4.29"
+    "vue": "^3.4.29",
+    "vue-router": "^4.0.13"
   },
   "devDependencies": {
     "@vitejs/plugin-vue": "^5.0.5",
